# frontend/src/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "AgentSettingsInput": {
        "properties": {
          "option_agent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/OptionAgentSettingsInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "option_postmortem": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentCommandSettingsInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "option_thesis": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentCommandSettingsInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "thesis_monitor": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ThesisMonitorSettingsInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "AgentSettingsInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RefreshJobResponse": {
        "additionalProperties": true,
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "job_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Name"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "summary": {
            "title": "Summary"
          }
        },
        "title": "RefreshJobResponse",
        "type": "object"
      },
...
      "ResearchSourcesInput": {
        "properties": {
          "blogs": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchBlogsSettingsInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "news": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchNewsSettingsInput"
              },
              {
                "type": "null"
              }
            ]
          },
          "x": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResearchXSettingsInput"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "ResearchSourcesInput",
        "type": "object"
      },
...
      "SettingsResponse": {
        "properties": {
          "agents": {
            "additionalProperties": true,
            "title": "Agents",
            "type": "object"
          },
          "config": {
            "additionalProperties": true,
            "title": "Config",
            "type": "object"
          },
          "integration": {
            "additionalProperties": true,
            "title": "Integration",
            "type": "object"
          },
          "sources": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceAuditResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/ApiStatusResponse"
          }
        },
        "required": [
          "status"
        ],
        "title": "SettingsResponse",
        "type": "object"
      },
...
      "ThesisAutomationInput": {
        "properties": {
          "dry_run": {
            "default": false,
            "title": "Dry Run",
            "type": "boolean"
          },
          "force": {
            "default": false,
            "title": "Force",
            "type": "boolean"
          },
          "symbols": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbols"
          }
        },
        "title": "ThesisAutomationInput",
        "type": "object"
      },
...
      "ThesisHistoryResponse": {
        "additionalProperties": true,
        "properties": {
          "review_events": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Review Events",
            "type": "array"
          },
          "revisions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Revisions",
            "type": "array"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          }
        },
        "title": "ThesisHistoryResponse",
        "type": "object"
      },
      "ThesisInput": {
        "properties": {
          "automation_policy": {
            "anyOf": [
              {
                "enum": [
                  "auto",
                  "manual_lock"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Automation Policy"
          },
          "catalyst": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalyst"
          },
          "catalysts": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Catalysts"
          },
          "change_rationale": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Rationale"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "conviction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conviction"
          },
          "direction": {
            "anyOf": [
              {
                "enum": [
                  "bullish",
                  "bearish"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Direction"
          },
          "evidence_coverage_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Coverage Status"
          },
          "evidence_links": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Links"
          },
          "horizon_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Horizon Date"
          },
          "invalidation": {
            "default": "",
            "title": "Invalidation",
            "type": "string"
          },
          "invalidation_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidation Price"
          },
          "invalidation_rules": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Invalidation Rules"
          },
          "lifecycle_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lifecycle Status"
          },
          "max_loss": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Loss"
          },
          "next_review_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Review Date"
          },
          "pillars": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pillars"
          },
          "review_cadence_days": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Cadence Days"
          },
          "scenarios": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenarios"
          },
          "schema_version": {
            "default": 3,
            "enum": [
              1,
              2,
              3
            ],
            "title": "Schema Version",
            "type": "integer"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "thesis": {
            "title": "Thesis",
            "type": "string"
          },
          "timeframe": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timeframe"
          },
          "why": {
            "default": "",
            "title": "Why",
            "type": "string"
          }
        },
        "required": [
          "thesis"
        ],
        "title": "ThesisInput",
        "type": "object"
      },
...
      "ThesisMutationResponse": {
        "additionalProperties": true,
        "properties": {
          "thesis": {
            "additionalProperties": true,
            "title": "Thesis",
            "type": "object"
          },
          "thesis_monitor": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis Monitor"
          }
        },
        "required": [
          "thesis"
        ],
        "title": "ThesisMutationResponse",
        "type": "object"
      },
      "ThesisReviewInput": {
        "properties": {
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          },
          "outcome": {
            "default": "unchanged",
            "enum": [
              "unchanged",
              "updated",
              "invalidated",
              "closed"
            ],
            "title": "Outcome",
            "type": "string"
          },
          "reviewed_evidence_cutoff": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed Evidence Cutoff"
          }
        },
        "title": "ThesisReviewInput",
        "type": "object"
      },
      "ThesisReviewResponse": {
        "additionalProperties": true,
        "properties": {
          "review": {
            "additionalProperties": true,
            "title": "Review",
            "type": "object"
          },
          "thesis_monitor": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thesis Monitor"
          }
        },
        "required": [
          "review"
        ],
        "title": "ThesisReviewResponse",
        "type": "object"
      },
      "TickerDecisionSnapshotResponse": {
        "additionalProperties": true,
        "properties": {
          "found": {
            "default": true,
            "title": "Found",
            "type": "boolean"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          }
        },
        "required": [
          "symbol"
        ],
        "title": "TickerDecisionSnapshotResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Personal Investment Panel",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/refresh-jobs/{job_name}": {
      "post": {
        "operationId": "launch_refresh_job_api_refresh_jobs__job_name__post",
        "parameters": [
          {
            "in": "path",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Launch Refresh Job"
      }
    },
    "/api/refresh-jobs/{job_name}/background": {
      "post": {
        "operationId": "launch_refresh_job_background_api_refresh_jobs__job_name__background_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_name",
            "required": true,
            "schema": {
              "title": "Job Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshJobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Launch Refresh Job Background"
      }
    },
...
    "/api/settings/agents": {
      "patch": {
        "operationId": "update_agent_settings_api_settings_agents_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSettingsInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Agent Settings"
      }
    },
    "/api/settings/research-sources": {
      "patch": {
        "operationId": "update_research_sources_api_settings_research_sources_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchSourcesInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Research Sources"
      }
    },
...
    "/api/theses/{symbol}": {
      "put": {
        "operationId": "save_thesis_endpoint_api_theses__symbol__put",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThesisInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisMutationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Save Thesis Endpoint"
      }
    },
    "/api/theses/{symbol}/history": {
      "get": {
        "operationId": "thesis_history_endpoint_api_theses__symbol__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Thesis History Endpoint"
      }
    },
    "/api/theses/{symbol}/review": {
      "post": {
        "operationId": "review_thesis_endpoint_api_theses__symbol__review_post",
        "parameters": [
          {
            "in": "path",
            "name": "symbol",
            "required": true,
            "schema": {
              "title": "Symbol",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ThesisReviewInput"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Review Thesis Endpoint"
      }
    },
    "/api/thesis-monitor/automation": {
      "post": {
        "operationId": "run_thesis_automation_endpoint_api_thesis_monitor_automation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ThesisAutomationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ThesisAutomationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Thesis Automation Endpoint"
      }
    },
    "/api/tickers/{ticker}": {
      "get": {
        "operationId": "ticker_detail_api_tickers__ticker__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ticker Detail"
      }
    },
    "/api/tickers/{ticker}/decision-snapshot": {
      "get": {
        "operationId": "ticker_decision_snapshot_api_tickers__ticker__decision_snapshot_get",
        "parameters": [
          {
            "in": "path",
            "name": "ticker",
            "required": true,
            "schema": {
              "title": "Ticker",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerDecisionSnapshotResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ticker Decision Snapshot"
      }
    },
...
  }
}